requirements.txt
flask>=3.1.1
werkzeug>=3.1.3
pymongo>=4.15.5
resend>=2.29.0
python-dotenv>=1.2.1
requests>=2.32.4
email-validator>=2.2.0
gunicorn>=23.0.0
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.42
flask-cors>=4.0.0
bson
flask-cors
